refactor: Reestructure the code so Named Entity Recognition has its own folder.
