+++ lib/ner/index.js
@@ -21,44 +21,10 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-const { Classifier, LogisticRegressionClassifier } = require('./classifiers');
-const { Language } = require('./language');
-const { Mathops, Vector, Matrix } = require('./math');
-const { NlgManager } = require('./nlg');
-const {
-  NlpClassifier,
-  NlpManager,
-  NlpUtil,
-  NlpExcelReader,
-} = require('./nlp');
-const { NerManager, NamedEntity } = require('./ner');
-const { SentimentAnalyzer, SentimentManager } = require('./sentiment');
-const { Evaluator, SimilarSearch } = require('./util');
-const { XTableUtils, XTable, XDoc } = require('./xtables');
-const { ConversationContext, MemoryConversationContext, Recognizer } = require('./recognizer');
+const NerManager = require('./ner-manager');
+const NamedEntity = require('./named-entity');
 
 module.exports = {
-  Classifier,
-  LogisticRegressionClassifier,
-  Language,
-  Mathops,
-  Vector,
-  Matrix,
-  NlgManager,
   NerManager,
-  NlpClassifier,
-  NlpManager,
-  NlpUtil,
-  NlpExcelReader,
   NamedEntity,
-  SentimentAnalyzer,
-  SentimentManager,
-  Evaluator,
-  SimilarSearch,
-  XTableUtils,
-  XTable,
-  XDoc,
-  ConversationContext,
-  MemoryConversationContext,
-  Recognizer,
 };
