+++ lib/ner/ner-manager.js
