+++ lib/nlp/index.js
@@ -21,18 +21,14 @@
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-const NerManager = require('./ner-manager');
 const NlpUtil = require('./nlp-util');
 const NlpClassifier = require('./nlp-classifier');
 const NlpManager = require('./nlp-manager');
-const NamedEntity = require('./named-entity');
 const NlpExcelReader = require('./nlp-excel-reader');
 
 module.exports = {
-  NerManager,
   NlpUtil,
   NlpClassifier,
   NlpManager,
-  NamedEntity,
   NlpExcelReader,
 };
