+++ lib/nlp/nlp-manager.js
@@ -24,8 +24,7 @@
 const fs = require('fs');
 const Handlebars = require('handlebars');
 const { Language } = require('../language');
-const NerManager = require('./ner-manager');
-const NamedEntity = require('./named-entity');
+const { NerManager, NamedEntity } = require('../ner');
 const { SentimentManager } = require('../sentiment');
 const NlpUtil = require('./nlp-util');
 const NlpClassifier = require('./nlp-classifier');
